Minor changes and bug fixes
Changelog:
- Added option to always have "pre" checkbox always checked on (or off). To do this, add this line to your base/config.ini:
always_pre = true
- No more need to put the tags in courtroom_config.ini in specific order.
- Added mute list and line edit stylesheets. The Line Edit one is used in AONotePicker QLineEdit object.
You can simply add them by adding the proper tags to your courtroom_config.ini file in the theme folder:
[MUTE LIST]
QListView {
}
[LINE EDIT]
QLineEdit {
}

and editing as such.
