+++ aoapplication.h
@@ -178,7 +178,7 @@ public:
   QString read_char_ini(QString p_char, QString p_search_line, QString target_tag, QString terminator_tag);
 
   //Returns the text between target_tag and terminator_tag in courtroom_config.ini
-  QString get_stylesheet(QString target_tag, QString terminator_tag);
+  QString get_stylesheet(QString target_tag);
 
   //Returns the side of the p_char character from that characters ini file
   QString get_char_side(QString p_char);
