+++ aonotepicker.cpp
@@ -1,5 +1,7 @@
 #include "aonotepicker.hpp"
 
+#include "courtroom.h"
+
 #include <QFileDialog>
 #include <QDebug>
 
@@ -7,18 +9,46 @@ AONotePicker::AONotePicker(QWidget *p_parent, AOApplication *p_ao_app) : QLabel(
 {
   m_parent = p_parent;
   ao_app = p_ao_app;
-
-//  QLineEdit *m_line = new QLineEdit(this);
-//  AOButton *m_button = new AOButton(this, ao_app);
-//  QHBoxLayout *m_layout = new QHBoxLayout(this);
-
-//  m_layout->addWidget(m_line);
-//  m_layout->addWidget(m_button);
-//  m_button->hide();
-//  this->setLayout(m_layout);
-
-//  connect(m_button, SIGNAL(clicked(bool)), this, SLOT(on_button_clicked()));
 }
 
 AONotePicker::~AONotePicker()
 {}
+
+void Courtroom::on_file_selected()
+{
+  for(int i=0; i < ui_note_area->m_layout->count() -1; ++i)
+    {
+      AONotePicker *f_notepicker = static_cast<AONotePicker*>(ui_note_area->m_layout->itemAt(i)->widget());
+      f_notepicker->m_hover->set_image("note_select.png");
+    }
+
+  AOButton *f_button = static_cast<AOButton*>(sender());
+  AONotePicker *f_notepicker = static_cast<AONotePicker*>(f_button->parent());
+  current_file = f_notepicker->real_file;
+  load_note();
+  f_button->set_image("note_select_selected.png");
+}
+
+
+void Courtroom::on_set_file_button_clicked()
+{
+  AOButton *f_button = static_cast<AOButton*>(sender());
+  AONotePicker *f_notepicker = static_cast<AONotePicker*>(f_button->parent());
+  QString f_filename = QFileDialog::getOpenFileName(this, "Open File");
+  if(f_filename != "")
+    {
+      f_notepicker->m_line->setText(f_filename);
+      f_notepicker->real_file = f_filename;
+
+      set_note_files();
+    }
+}
+
+void Courtroom::on_delete_button_clicked()
+{
+  AOButton *f_button = static_cast<AOButton*>(sender());
+  AONotePicker *f_notepicker = static_cast<AONotePicker*>(f_button->parent());
+  ui_note_area->m_layout->removeWidget(f_notepicker);
+  delete f_notepicker;
+  set_note_files();
+}
