+++ text_file_functions.cpp
@@ -430,7 +430,7 @@ QString AOApplication::get_font_name(QString p_identifier, QString p_file)
     return f_result;
 }
 
-QString AOApplication::get_stylesheet(QString target_tag, QString terminator_tag)
+QString AOApplication::get_stylesheet(QString target_tag)
 {
   QString design_ini_path = get_theme_path() + "courtroom_config.ini";
 
@@ -451,16 +451,18 @@ QString AOApplication::get_stylesheet(QString target_tag, QString terminator_tag
   {
     QString line = in.readLine();
 
-    if (QString::compare(line, terminator_tag, Qt::CaseInsensitive) == 0)
-      break;
-
     if (line.startsWith(target_tag, Qt::CaseInsensitive))
     {
       tag_found = true;
       continue;
     }
 
-    if(tag_found) f_text.append(line);
+    if(tag_found)
+      {
+        if((line.startsWith("[") && line.endsWith("]")))
+           break;
+        f_text.append(line);
+      }
   }
 
   design_ini.close();
